Resolve cyclic dependencies
