+++ compiler/BIRReader.h
@@ -1,10 +1,18 @@
 #ifndef BIRREADER_H
 #define BIRREADER_H
 
-#include "BIR.h"
+#include "BalFuncParam.h"
+#include "BalFunction.h"
 #include "BalInvokableType.h"
-#include "BalLocation.h"
+#include "BalMoveInsn.h"
+#include "BalNonTerminatorInsn.h"
+#include "BalOperand.h"
+#include "BalParam.h"
+#include "BalTerminatorInsn.h"
 #include "BalTypeDecl.h"
+#include "BalAbstractInsn.h"
+#include "BalBasicBlock.h"
+#include "BIR.h"
 #include <fstream>
 
 class ConstantPoolSet;
