+++ compiler/codegen/BalTerminatorInsn.cpp
@@ -5,8 +5,6 @@ TerminatorInsn::TerminatorInsn(Location *pos, InstructionKind kind,
                                Operand *lOp, BIRBasicBlock *then)
     : AbstractInsn(pos, kind, lOp), thenBB(then) {}
 
-void TerminatorInsn::translate(__attribute__((unused)) LLVMModuleRef &modRef) {}
-
 BIRBasicBlock *TerminatorInsn::getNextBB() { return thenBB; }
 bool TerminatorInsn::getPatchStatus() { return patchRequire; }
 
@@ -14,5 +12,7 @@ void TerminatorInsn::setNextBB(BIRBasicBlock *block) { thenBB = block; }
 void TerminatorInsn::setPatchStatus(bool patchrequire) {
   patchRequire = patchrequire;
 }
+
+void TerminatorInsn::translate(__attribute__((unused)) LLVMModuleRef &modRef) {}
 
 } // namespace nballerina
